Add files via upload
PCSD3 clustering algorithm is specifically designed to cluster survey data that contains ordinal response variables (or Likert type response scale) along with a nominal variable specifying the categories of respondents. PCSD3 clusters the data based on the marking  patterns. The method does not require parameters  to be specified by the users. The users are required to provide data in specified format (see dataset and its description). PCSD3 recognize the respondent categories and identify their distinguised features and clusters data accordingly. Therefore the number of PCSD3 clusters are always equal to the number of designated categories of the respondents. PCSD3 utilize newly developed "Frequency based Probabilistic Scoring" and a dedicated "Decision Making procedure" for clustering.